+++ packages/eslint-config/eslintrc.json
@@ -40,7 +40,7 @@
     "@groww-tech/internal/two-line-between-class-members": "warn",
     "@groww-tech/internal/no-gcs-bucket-url": "error",
     "@groww-tech/internal/no-primitive-color-variables": "error",
-    "@groww-tech/internal/no-redeclare-primitive-variables": "error",
+    "@groww-tech/internal/no-redeclare-mint-tokens": "error",
     "quotes": [
       "warn",
       "single"
