+++ packages/eslint-plugin-internal/lib/rules/no-redeclare-mint-tokens.js
